Read username from phue.conf in order to know if we're authenticated
