+++ application/senic/nuimo_hub/config.py
@@ -14,6 +14,7 @@ log = get_logger(__name__)
 default_settings = dict(
     wifi_networks_path='/tmp/wifi_networks.json',
     wifi_setup_flag_path='/srv/nuimo_hub/data/WIFI_SETUP_REQUIRED',
+    hass_phue_config_path='/srv/nuimo_hass/data/phue.conf',
 )
